# backend/docs/swagger.yaml
definitions:
  handlers.AccountResponse:
    properties:
      description:
        example: Общий счёт для домашних расходов
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Семейный бюджет
        type: string
      owner_id:
        example: 42
        type: integer
    type: object
...
  handlers.ChangeRoleRequest:
    properties:
      role:
        enum:
        - viewer
        - editor
        - admin
        example: admin
        type: string
    required:
    - role
    type: object
  handlers.CreateAccountRequest:
    properties:
      description:
        example: Общий счёт для домашних расходов
        type: string
      name:
        example: Семейный бюджет
        type: string
    required:
    - name
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
  handlers.IDResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  handlers.InviteMemberRequest:
    properties:
      email:
        example: newmember@example.com
        type: string
      role:
        enum:
        - viewer
        - editor
        - admin
        example: editor
        type: string
    required:
    - email
    - role
    type: object
...
  handlers.MessageResponse:
    properties:
      message:
        example: password changed successfully
        type: string
    type: object
...
info:
  contact: {}
paths:
  /accounts:
    post:
      consumes:
      - application/json
      description: Создаёт новый счёт для управления финансами. Создатель счёта автоматически
        получает роль Owner и может приглашать других участников, управлять их ролями
        и удалять счёт. Название счёта должно быть уникальным в рамках пользователя.
        Описание опционально.
      parameters:
      - description: Данные нового счёта. Название обязательно, описание опционально.
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Счёт успешно создан. Возвращается ID нового счёта.
          schema:
            $ref: '#/definitions/handlers.IDResponse'
        "400":
          description: Неверный формат данных. Проверьте наличие названия счёта.
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при создании счёта
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание нового счёта
      tags:
      - accounts
  /accounts/{id}:
    delete:
      description: 'Полностью удаляет счёт вместе со всеми связанными данными: транзакциями
        и участниками. Операция необратима. Доступна только владельцу счёта (роль
        Owner). После удаления все участники теряют доступ к счёту, и все транзакции
        становятся недоступны.'
      parameters:
      - description: ID счёта для удаления
        example: 1
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Счёт успешно удалён вместе со всеми данными
        "400":
          description: Неверный формат ID счёта
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Недостаточно прав. Удалять счёт может только его владелец (Owner)
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Счёт с указанным ID не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при удалении счёта
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление счёта
      tags:
      - accounts
    get:
      description: 'Возвращает детальную информацию о счёте по его ID: название, описание,
        владельца. Доступно всем участникам счёта. Для просмотра счёта пользователь
        должен быть его участником с любой ролью (Viewer, Editor, Admin, Owner).'
      parameters:
      - description: ID счёта
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Информация о счёте успешно получена
          schema:
            $ref: '#/definitions/handlers.AccountResponse'
        "400":
          description: Неверный формат ID счёта
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Пользователь не является участником данного счёта
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Счёт с указанным ID не найден
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение информации о счёте
      tags:
      - accounts
  /accounts/{id}/members:
    post:
      consumes:
      - application/json
      description: 'Добавляет нового участника в счёт с указанной ролью. Доступно
        только владельцу счёта (Owner). Приглашаемый пользователь должен быть зарегистрирован
        в системе. Роли: viewer (только просмотр), editor (создание/редактирование
        своих транзакций), admin (полные права на транзакции).'
      parameters:
      - description: ID счёта
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Email пользователя и его роль в счёте
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.InviteMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Участник успешно добавлен в счёт
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Неверный формат данных или неподдерживаемая роль
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Недостаточно прав. Приглашать участников может только Owner
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Пользователь с указанным email не найден в системе
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при добавлении участника
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Приглашение участника в счёт
      tags:
      - members
  /accounts/{id}/members/{user_id}:
    delete:
      description: Удаляет участника из счёта, лишая его доступа ко всем данным счёта.
        Доступно только владельцу счёта (Owner). Нельзя удалить самого владельца.
        После удаления участник теряет доступ к просмотру и редактированию транзакций.
        Созданные им транзакции остаются в счёте.
      parameters:
      - description: ID счёта
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: ID пользователя для удаления из счёта
        example: 42
        in: path
        name: user_id
        required: true
        type: integer
      responses:
        "204":
          description: Участник успешно удалён из счёта
        "400":
          description: Неверный формат ID счёта или пользователя
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Недостаточно прав. Удалять участников может только Owner
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Участник не найден в данном счёте
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при удалении участника
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удаление участника из счёта
      tags:
      - members
    patch:
      consumes:
      - application/json
      description: 'Изменяет роль существующего участника счёта. Доступно только владельцу
        счёта (Owner). Нельзя изменить роль самого владельца. Роли: viewer (только
        просмотр), editor (создание/редактирование своих транзакций), admin (полные
        права на транзакции). Изменение роли применяется немедленно.'
      parameters:
      - description: ID счёта
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: ID пользователя, чью роль нужно изменить
        example: 42
        in: path
        name: user_id
        required: true
        type: integer
      - description: Новая роль участника
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangeRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Роль участника успешно изменена
          schema:
            $ref: '#/definitions/handlers.MessageResponse'
        "400":
          description: Неверный формат данных или неподдерживаемая роль
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Отсутствует или невалидный JWT токен
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Недостаточно прав. Изменять роли может только Owner
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Участник не найден в данном счёте
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при изменении роли
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Изменение роли участника
      tags:
      - members
...
swagger: "2.0"